requirements.txt
streamlit
beautifulsoup4
openai
pandas
facebook-sdk
requests
instagrapi
python-dotenv